+++ CHANGELOG.md
@@ -1,7 +1,30 @@
 ### BioBlend v
 
+* Added support for Galaxy release 24.1.
+
+### BioBlend v1.3.0 - 2024-05-12
+
 * Dropped support for Python 3.7. Added support for Python 3.12. Added support
   for Galaxy releases 23.2 and 24.0.
+
+* Added ``copy_elements`` parameter to
+  ``HistoryClient.create_dataset_collection()`` and BioBlend.objects
+  ``History.create_dataset_collection()`` methods.
+
+* Added ``wait`` parameter to ``HistoryClient.delete_dataset()`` and
+  BioBlend.objects ``HistoryDatasetAssociation.delete()`` methods.
+
+* Added ``create_time_min``, ``create_time_max``, ``update_time_min``,
+  ``update_time_max``, ``view``, ``keys``, ``limit`` and ``offset`` parameters
+  to ``HistoryClient.get_histories()`` (thanks to
+  [cat-bro](https://github.com/cat-bro)).
+
+* Added ``create_time_min``, ``create_time_max``, ``update_time_min`` and
+  ``update_time_max`` parameters to ``HistoryClient.get_published_histories()``
+  (thanks to [cat-bro](https://github.com/cat-bro)).
+
+* Added ``keys`` parameter to ``HistoryClient.show_history()`` (thanks to
+  [cat-bro](https://github.com/cat-bro)).
 
 * Dropped broken ``deleted`` parameter of ``DatasetClient.show_dataset()``.
 
@@ -12,7 +35,16 @@
 * Classes defined in ``bioblend.galaxy.objects.wrappers`` are no more
   re-exported by ``bioblend.galaxy.objects``.
 
+* ``DatasetTimeoutException`` and ``DatasetCollectionTimeoutException`` are now
+  aliases for ``TimeoutException`` instead of subclasses.
+
 * Added support for the new "cancelling" invocation state.
+
+* Fixed ``InvocationClient.get_invocation_biocompute_object()`` method on
+  upcoming Galaxy 24.1 .
+
+* Improvements to linting and tests (thanks to
+  [Matthias Bernt](https://github.com/bernt-matthias)).
 
 ### BioBlend v1.2.0 - 2023-06-30